refactor(app): list renders
