+++ package.json
@@ -47,7 +47,7 @@
 		"@semantic-release/npm": "^10.0.3",
 		"@semantic-release/release-notes-generator": "^11.0.2",
 		"@types/node": "20.2.5",
-		"@types/react": "18.2.7",
+		"@types/react": "18.2.8",
 		"@types/react-dom": "18.2.4",
 		"@typescript-eslint/eslint-plugin": "^5.59.8",
 		"@typescript-eslint/parser": "^5.59.8",
@@ -79,6 +79,6 @@
 		"stylelint": "^15.6.2",
 		"stylelint-config-recommended": "^12.0.0",
 		"stylelint-config-tailwindcss": "^0.0.7",
-		"typescript": "5.0.4"
+		"typescript": "5.1.3"
 	}
 }
